Remove or rename mbuf fields lacking in-tree consumers or design doc
These fields were part of the broader project to clarify ownership of
parts of the mbuf throughout the stack and explore changes to the driver
and stack interface, and packet handoff.

Unfortunately these arrived in HEAD prior to the branchpoint for 10.0,
without examples of in-tree consumers or design documentation.  Further
changes to the model may arrive in HEAD after testing and review in a
project branch.

Thus, in 10 remove or mark them as reserved, while preserving the KBI,
to avoid specifying a kernel interface that does not yet exist.