Update the types used for fpcr and fpsr.
The special registers need to be accessed using a 64-bit general purpose
register, even if the special register is 32-bits.

Clang 3.6 adds a check to make sure the type and general purpose register
sizes line up. To fix this we need to use the correct sized type to get
the correct register size.