Don't strip paths from .so ldscript
lld does not have built-in search paths (e.g. /lib, /usr/lib) and
relies on -L arguments passed by the caller. As the linker is nearly
always invoked from the clang driver this is fine other than the
fact that /usr/lib/libc.so is an ldscript that refers to libc.so.7
which is in /lib, not /usr/lib.

Address this by leaving the paths in the GROUP entry rather than
stripping them out.