elfcopy: handle nameless global symbols
Global symbol names are stored in a separate buffer and appended to the
local symbol names when creating .strtab.

Prior to this change both local and global nameless symbol names were
stored with st_name = 0.  This was correct for local symbols; the string
table has an explicit '\0' at the beginning.  However, the global symbol
list starts with a named symbol.  Nameless global symbols incorrectly
took on that name.

Now we store -1 in st_name for nameless global symbols, and update them
to 0 while finalizing the global string offsets.