sys/boot/i386: skip as lld fails to link it
lld does not currently support -N. It is not yet clear if there's enough
value in supporting -N in lld, or if we should adjust our build to avoid
it. It has also been suggested that it is not necessary for some of the
i386 boot components.

llvm.org/pr23214