uathload: skip building, it does not build with lld
The uathload Makefile uses ld -r to convert a binary blob into an ELF
object. lld does not currently supprt this use, and it's unclear if
this is a sufficiently common use case to warrant adding support for
it in lld.

I suggest replacing this use with objcopy (once we've finalized the
switch to the ELF Tool Chain version).