libc: add per-arch backwards compatability symbol control
The FreeBSD/arm ABI changed in FreeBSD 10 and the previous one is no
longer supported as of FreeBSD 11. FreeBSD/arm64 and FreeBSD/riscv
have no releases prior to FreeBSD 11. Do not enable symver backwards
compatabilitiy for earlier FreeBSD releases.

FreeBSD 7 is the minimum LIB_MIN_COMPAT as it is the one that introduced
libc.so.7.

Do not enable any symver backwards compatability when building
WITHOUT_SYMVER.

PR:		207712
Sponsored by:	The FreeBSD Foundation
Differential Revision: https://reviews.freebsd.org/D5650