Disconnect objdump from the build
The in-tree binutils is old and will not be updated.  Other tools are
migrating to copyfree alternatives but there is not yet a direct
replacement for objdump.  Just remove it as it is not used by the build
and updated, maintained versions are available from ports and packages.

TODO: Add to ObsoleteFiles.inc