arm64 make-memstick.sh: use `set -e` to abort if any step fails
Also remove the now-redundant error handling that was only for makefs.

This script was run on an older FreeBSD host that lacked efi-on-mbr
support in makefs. A warning was emitted on the console (from makefs)
but the script continued running and exited with 0.

Approved by:	re (not really)
Sponsored by:	The FreeBSD Foundation