Always pass -m to ld for converting binary files to kernel ELF objects
This is in preparation for linking with LLVM's LLD linker, which does
not have a compiled-in default output emulation. LLD requires that the
emulation be specified via the -m option, or obtained from the object
file(s) being linked.

This also allos the same LLD linker binary to be used to build all
targets.

Sponsored by:	The FreeBSD Foundation
Differential Revision:	https://reviews.freebsd.org/D7837