add octeon+ as an alias for octeon in GCC & binutils
In r208737 jmallett@ added support for the "mips64r2" architecture
and "octeon" CPU. Upstream binutils has in addition the "octeon+" CPU,
and the special saa/saad instructions are only available in octeon+, not
octeon.

This allows the use of octeon+ in kernel config files, for use with both
external tool chain and in-tree GCC/binutils.

PR:		216516