Add NetBSD makefs TODO and git log
This is a partial 'git log' of NetBSD's usr.sbin/makefs.

On Nov 7 2010 cognet@ sync'd with NetBSD. The last NetBSD commit prior
to that was:

commit 29d74a517fe40cbc99df3b95d4618d58e9c05df7
Author:	christos <christos>
Date:	Wed Oct 27 18:51:34 2010 +0000

    Handle > 2GB images. Thanks to Greg Troxel for testing!

Commit log entries for later NetBSD changes that already exist in
FreeBSD, and for filesystems not in FreeBSD, have also been removed.
For filessystems not yet in FreeBSD like msdos I assume that we will
take a snapshot of the latest version of NetBSD's code and won't
be concerned with the interim changes.