Disable binutils everywhere except amd64/i386 (for GNU as)
As of June 2017 in stock FreeBSD we use three programs from binutils:
as, ld, and objdump. Objdump is not required by the build and can just
be skipped. For the linker we're migrating to a combination of in-tree
lld and external binutils port or package, depending on arch.  That
leaves the assembler, as. We're generally planning to migrate to Clang's
integrated assembler (IAS), but there are a small number of asm files on
both i386 and amd64 that cannot be assembled by IAS, so leave binutils
enabled for those while working on a path forward.