libgcc_s: Make compiler-rt's quad-math routines available to libgcc_s
compiler-rt includes 128-bit floating point routines, but we previously
did not export them in our compiler-rt-based libgcc_s.  These routines
are required by Fortran programs compiled with GCC 4.6 or later.

Additional work is required to enable this for amd64, but this should
have an effect on arm64 and riscv64.

Sponsored by:	The FreeBSD Foundation
Differential Revision:	https://reviews.freebsd.org/D11482