lld: fix crash linking the FreeBSD/arm64 kernel
As reported by dim@ in LLVM PR33821 lld was crashing while linking the
FreeBSD/arm64 kernel, with a null linker script command section:

  1337      Cmd->Sec->addSection(Sentinel);
  (gdb) p Cmd->Sec
  $4 = (lld::elf::OutputSection *) 0x0

Under discussion upstream, https://llvm.org/pr33821.  Not yet certain if
this fix is correct, but it does prevent the crash and the LLD test
suite passes with the change.