Use conventional test for retpoline option and compiler feature
Previously we used .error to abort if the RETPOLINE option was enabled
but not listed in COMPILER_FEATURES.  However this would also apply at
installkernel time, preventing the installation of a retpoline-enabled
kernel if the existing system compiler did not support retpoline. Just
use the standard style of test to enable retpoline.

Sponsored by:	The FreeBSD Foundation