Correct typo in ATA_WRITE_UNCORRECTABLE_PSEUDO
This file has an Intel copyright but I can find no evidence that it
exists outside of FreeBSD.  Assume that FreeBSD is the reference
implementation and correct _PSUEDO to _PSEUDO before this ends up
elsehwere.

Also correct a typo in the comment for these values.

Reviewed by:	jimharris
Sponsored by:	The FreeBSD Foundation
Differential Revision:	https://reviews.freebsd.org/D3715