pwd_mkdb: retire legacy v3 db support (-l option)
pwd_mkdb has emitted v4 password database records since 2003 (r113596)
in addition to v3, and as of r283981 by default it emitted only v4.
As described in r283981, retire the -l legacy option.

Additionally, set the db hash to be in network byte order (big-endian)
as with other integers in the database.  The db(3) functions work with a
hash of either endianness.  Using a fixed endianness allows us to create
a password database that's reproducible regardless of host endianness.

The -B and -L options were originally added to set the endianness of v3
records emitted by pwd_mkdb, but they also set the db hash endiannes and
so have been retained temporarily.

Sponsored by:	The FreeBSD Foundation
Differential Revision:	https://reviews.freebsd.org/D15144