Error out on attempt to link amd64 with binutils 2.17.50 linker
As of r333461 we require ifunc support to link the amd64 kernel.
The default in-tree bootstrap linker is lld and has the required
support, as does any modern out-of-tree binutils linker.  The
essentially obsolete in-tree BFD linker does not, so produce an error
rather than a broken kernel.

Sponsored by:	The FreeBSD Foundation
Differential Revision:	https://reviews.freebsd.org/D15378