if_muge: Use lock assertion instead of broken locking in lan78xx_chip_init
lan78xx_chip_init locked the driver's mutex if not already locked, but
unlocked it only in the case of error.  This gives a pretty clear
indication that the function is being called with the lock already held,
so just replace it with a lock assertion.

Sponsored by:	The FreeBSD Foundation