sqlite: do not return error if filesystem does not implement posix_fallocate
posix_fallocate returns EINVAL if the operation is not supported by the
filesystem.  This may be encountered on copy-on-write filesystems such
as ZFS.

It is unclear to me why POSIX settled on EINVAL for this.

In any case, we can treat posix_fallocate as advisory and just carry on
if not supported by the filesystem.

Reference: https://twitter.com/DRichardHipp/status/1002919969510821888