linux64: add arm64 build details
The arm64 linuxulator needs different arguments for the objcopy
invocation used to build the linux VDSO.

The VDSO format and build is inherently arch- and OS-dependent, and
on some architectures is even not an ELF object.  I have opted for some
duplication (i.e., repeating the full command in .if blocks) for
simplicity.  This ought to be revisited upon the addition of a third
64-bit Linuxulator arch.

Sponsored by:	Turing Robotic Industries