Revert "elfcopy: delete filter_reloc, it is broken and unnecessary"
This reverts commit fbeae6e8c2dbb0367a611fdc78da43c8cae5ed17.

The reverted commit fixed the case of stripping statically linked
binaries with relocations (which may come from ifunc use, for example).
As a side effect this changed the behaviour when stripping object
files - the output was broken both before and after rS339350, in
different ways.

Unfortunately GCC's build process relies on the previous behaviour, so
first restore filter_reloc so that it can be fixed.