Implement a BSD licensed crtbegin/crtend
These are needed for .ctors/.dtors and .jcr handling. The former needs
all the function pointers to be called in the correct order from the
.init/.fini section. The latter just needs to call a gcj specific function
if it exists with a pointer to the start of the .jcr section.

Tested on amd64 and arm64, other architectures still need to be
implemented.

From D17587 diff 49573 adapted to wipbsd-12 (gnu/lib/Makefile and
src.opts.mk).

Submitted by:	andrew
Differential Revision:	https://reviews.freebsd.org/D17587