llvm-objdump: install as objdump on Clang-using architectures
Install LLVM's objdump as /usr/bin/objdump on those architecures which
use Clang as the system compiler, as one step in the retirement of the
obsolete binutils 2.17.50.

It supports many, but not all, of GNU objdump's command line options.
Users requiring the full functionality of GNU objdump should install
the binutils port.

PR:		229046
Differential Revision:	https://reviews.freebsd.org/D18307