linuxulator: rename .s files to .S
As part of the migration away from obsolete binutils we want to retire
GNU as.  Most assembly files used on amd64 have a .S extension and are
assembled with Clang's integrated assembler (via rules in
share/mk/bsd.suffixes.mk).

The Linuxulator has bespoke rules for invoking the compiler driver to
assemble .s files and was already using Clang's IAS, but rename the
assembly files for consistency.

Differential Revision:	https://reviews.freebsd.org/D18344